+++ src/main/java/ru/vyarus/dropwizard/guice/module/installer/bundle/GuiceyBootstrap.java
@@ -10,6 +10,7 @@ import ru.vyarus.dropwizard.guice.module.context.ConfigurationContext;
 import ru.vyarus.dropwizard.guice.module.context.option.Option;
 import ru.vyarus.dropwizard.guice.module.installer.FeatureInstaller;
 import ru.vyarus.dropwizard.guice.module.lifecycle.GuiceyLifecycleListener;
+import ru.vyarus.dropwizard.guice.module.yaml.YamlConfig;
 
 import java.util.Arrays;
 import java.util.List;
@@ -38,18 +39,10 @@ public class GuiceyBootstrap {
 
     private final ConfigurationContext context;
     private final List<GuiceyBundle> iterationBundles;
-    private final Configuration configuration;
-    private final Environment environment;
-    private final Bootstrap bootstrap;
 
-    public GuiceyBootstrap(final ConfigurationContext context, final List<GuiceyBundle> iterationBundles,
-                           final Configuration configuration, final Environment environment,
-                           final Bootstrap bootstrap) {
+    public GuiceyBootstrap(final ConfigurationContext context, final List<GuiceyBundle> iterationBundles) {
         this.context = context;
         this.iterationBundles = iterationBundles;
-        this.configuration = configuration;
-        this.environment = environment;
-        this.bootstrap = bootstrap;
     }
 
     /**
@@ -59,7 +52,7 @@ public class GuiceyBootstrap {
      * @return dropwizard bootstrap instance
      */
     public Bootstrap bootstrap() {
-        return bootstrap;
+        return context.getBootstrap();
     }
 
     /**
@@ -68,14 +61,98 @@ public class GuiceyBootstrap {
      */
     @SuppressWarnings("unchecked")
     public <T extends Configuration> T configuration() {
-        return (T) configuration;
+        return (T) context.getConfiguration();
+    }
+
+    /**
+     * May be used to access current configuration value by exact path. This is helpful for bundles universality:
+     * suppose bundle X requires configuration object XConf, which is configured somewhere inside application
+     * configuration. We can require configuration path in bundle constructor and use it to access required
+     * configuration object: {@code new X("sub.config.path")}.
+     *
+     * @param yamlPath target value yaml path
+     * @param <T> value type
+     * @return configuration value by path or null if value is null or path not exists
+     * @see #yamlConfig() for custom configuration searches
+     */
+    public <T> T configuration(final String yamlPath) {
+        return yamlConfig().valueByPath(yamlPath);
+    }
+
+    /**
+     * May be used to access unique sub configuration object. This is helpful for bundles universality:
+     * suppose bundle X requires configuration object XConf and we are sure that only one declaration of XConf would
+     * be used in target configuration class, then we can simply request it:
+     * {@code configuration(XConf.class) == <instance of XConf or null>}.
+     * <p>
+     * Note that uniqueness is checked by declaration class:
+     * <pre>{@code class Config extends Configuration {
+     *     Sub sub;
+     *     SubExt ext; // SubExt extends Sub
+     * }}</pre>
+     * are unique declarations (declaration of the same type never appears in configuration on any level).
+     * {@code configuration(Sub.class) == sub} and {@code configuration(SubExt.class) == ext}.
+     * <p>
+     * Example of accessing server config from dropwizard configuration:
+     * {@code configuration(ServerFactory.class) == DefaultServerFactory (or SimpleServerFactory)}
+     * (see dropwizard {@link Configuration} class).
+     *
+     * @param type target configuration declaration type
+     * @param <T> declaration type
+     * @param <K> required value type (may be the same or extending type)
+     * @return unique configuration value or null if value is null or no declaration found
+     * @see #yamlConfig() for custom configuration searches
+     */
+    public <T, K extends T> K configuration(final Class<T> type) {
+        return yamlConfig().valueByUniqueDeclaredType(type);
+    }
+
+    /**
+     * IMPORTANT: method semantic is different from {@link #configuration(Class)}, which use direct class
+     * declaration match, whereas this method searches by all assignable types.
+     * <pre>{@code class Config extends Configuration {
+     *     Sub sub;
+     *     SubExt ext; // SubExt extends Sub
+     * }}</pre>
+     * {@code configurations(Sub.class) == [sub, ext]}, but {@code configurations(SubExt.class) == [ext]}.
+     * <p>
+     * Useful when multiple sub configuration objects could be used and all of them are required in some
+     * universal bundle.
+     * <p>
+     * Note: only custom types may be used (sub configuration objects), not Integer, Boolean, List, etc.
+     *
+     * @param type target configuration type
+     * @param <T> value type
+     * @return list of configuration values with required type or empty list
+     * @see #yamlConfig() for custom configuration searches
+     */
+    public <T> List<? extends T> configurations(final Class<T> type) {
+        return yamlConfig().valuesByType(type);
+    }
+
+    /**
+     * Raw configuration introspection info. Could be used for more sophisticated configuration searches then
+     * provided in shortcut methods.
+     * <p>
+     * Note that configuration is analyzed using jackson serialization api, so not all configured properties
+     * could be visible (when property getter is not exists or field not annotated).
+     * <p>
+     * Returned object contains all resolved configuration paths. Any path element could be traversed like a tree.
+     * See find* and value* methods as an examples of how stored paths could be traversed.
+     *
+     * @return detailed configuration object
+     * @see ru.vyarus.dropwizard.guice.module.yaml.YamlConfigInspector for configuration introspection details
+     * @see ru.vyarus.dropwizard.guice.module.yaml.module.Config for available guice configuration bindings 
+     */
+    public YamlConfig yamlConfig() {
+        return context.getYamlConfig();
     }
 
     /**
      * @return environment instance
      */
     public Environment environment() {
-        return environment;
+        return context.getEnvironment();
     }
 
     /**
@@ -91,7 +168,7 @@ public class GuiceyBootstrap {
      * @return dropwizard application instance
      */
     public Application application() {
-        return bootstrap.getApplication();
+        return context.getBootstrap().getApplication();
     }
 
     /**